Move metrics owner to sandbox on x
