+++ ledger/sandbox-on-x/src/main/scala/com/daml/ledger/sandbox/MetricsOwner.scala
@@ -1,7 +1,4 @@
-// Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.daml.platform.metrics
+package com.daml.ledger.sandbox
 
 import java.util.concurrent.TimeUnit
